Use ints for userid parameters

+++ osu.Server.Spectator/Hubs/MultiplayerHub.cs
@@ -181,7 +181,7 @@ namespace osu.Server.Spectator.Hubs
             }
         }
 
-        public async Task TransferHost(long userId)
+        public async Task TransferHost(int userId)
         {
             var room = await getLocalUserRoom();
